Integrated Flight Computer/tools/kerbin_map/python/requirements.txt
krpc>=0.5.4
numpy>=1.24
Pillow>=10.0
